move over old tests, test max floating pt
