+++ RealmSwift/Tests/ObjectAccessorTests.swift
@@ -38,11 +38,15 @@ class ObjectAccessorTests: TestCase {
         XCTAssertEqual(object.floatCol, 20 as Float)
         object.floatCol = 20.2
         XCTAssertEqual(object.floatCol, 20.2 as Float)
+        object.floatCol = 16777217
+        XCTAssertEqual(Double(object.floatCol), 16777216.0 as Double)
 
         object.doubleCol = 20
         XCTAssertEqual(object.doubleCol, 20)
         object.doubleCol = 20.2
         XCTAssertEqual(object.doubleCol, 20.2)
+        object.doubleCol = 16777217
+        XCTAssertEqual(object.doubleCol, 16777217)
 
         object.stringCol = ""
         XCTAssertEqual(object.stringCol, "")
@@ -73,5 +77,58 @@ class ObjectAccessorTests: TestCase {
         Realm().create(SwiftObject.self)
         setAndTestAllProperties(object)
         Realm().commitWrite()
+    }
+
+    func testIntSizes() {
+        let realm = realmWithTestPath()
+
+        let v16 = Int16(1) << 12
+        let v32 = Int32(1) << 30
+        // 1 << 40 doesn't auto-promote to Int64 on 32-bit platforms
+        let v64 = Int64(1) << 40
+        realm.write {
+            let obj = SwiftAllIntSizesObject()
+
+            obj.int16 = v16
+            XCTAssertEqual(obj.int16, v16)
+            obj.int32 = v32
+            XCTAssertEqual(obj.int32, v32)
+            obj.int64 = v64
+            XCTAssertEqual(obj.int64, v64)
+
+            realm.add(obj)
+        }
+
+        let obj = realm.objects(SwiftAllIntSizesObject.self).first!
+        XCTAssertEqual(obj.int16, v16)
+        XCTAssertEqual(obj.int32, v32)
+        XCTAssertEqual(obj.int64, v64)
+    }
+
+    func testLongType() {
+        let longNumber = 17179869184
+        let intNumber = 2147483647
+        let negativeLongNumber = -17179869184
+        let updatedLongNumber = 8589934592
+
+        let realm = realmWithTestPath()
+
+        realm.beginWrite()
+        realm.create(SwiftIntObject.self, value: [longNumber])
+        realm.create(SwiftIntObject.self, value: [intNumber])
+        realm.create(SwiftIntObject.self, value: [negativeLongNumber])
+        realm.commitWrite()
+
+        let objects = realm.objects(SwiftIntObject.self)
+        XCTAssertEqual(objects.count, Int(3), "3 rows expected")
+        XCTAssertEqual(objects[0].intCol, longNumber, "2 ^ 34 expected")
+        XCTAssertEqual(objects[1].intCol, intNumber, "2 ^ 31 - 1 expected")
+        XCTAssertEqual(objects[2].intCol, negativeLongNumber, "-2 ^ 34 expected")
+
+        realm.beginWrite()
+        objects[0].intCol = updatedLongNumber
+        realm.commitWrite()
+
+        XCTAssertEqual(objects[0].intCol, updatedLongNumber, "After update: 2 ^ 33 expected")
     }
 }
